+++ src/lib.rs
@@ -11,6 +11,7 @@ pub mod fasta;
 pub mod intervals;
 pub mod math;
 pub mod metrics;
+pub mod overlapper;
 pub mod plotting;
 pub mod progress;
 pub mod sam;